Device memory initialization code & address calculation
Since CL lacks malloc or generally dynamic memory allocation, the only
way to use this is to allocate the buffers on the host and then hook the
pointers up on the device, hence the mem_hook_up() kernel.

It not necessary to calculate the addresses (indices) of the source
values for every iteration. It is sufficient to calculate them once in
advance and then store them on the device.

--HG--
branch : for_andi