Add additional template parameter for OpenCL Stepper
This template parameter specifies the type of the data which is to be
copied onto the Device. This facilitates some parts of the code (e.g.
the sizeof_data interface method can be removed) but most importantly it
is necessary to copy data back to the Host from the Device.
The reason for this is that OpenCL only proviedes raw (void *) access to
buffer memory which is impractical for copying single cells (array
indices).

--HG--
branch : for_andi