One Bugfix for Patternoptimizer and new OptimizerTest setup
The Optimizertest.h, PatternOptimizerTest.h and SimplexOptimizerTest are merged in Optimizer.h. Each Testfunction get a own Testsuite with
initialization of the SimulationParameters and initialization of the evaluation function in the setUp() function and the INFO output in the tearDown().

--HG--
rename : src/misc/test/unit/patternoptimizertest.h => src/misc/test/unit/optimizertest.h