Moved MPI-enabled tests to corresponding directory
If configured with WITH_MPI, some of our utility classes (e.g. Pipe)
will create an MPILayer, which is not necessarily used, but in an
invalid state since MPI is never initialized for a unit test. Moving
the tests has the drawback of not checking the components if MPI is disabled.