MPILayer will now auto-initialize Typemaps
This relieves the user from having to manually call
Typemaps::initializeMaps(). There is a catch though: now an MPILayer
may only be constructed if MPI was already initialized. That's fine
though because an MPILayer instance isn't functional before MPI
initialization. This change did prompt a lot of trivial changes in
unit tests.