unstructured: added proxy writer for SoA unstructured grids which sorts the output before writing it
In SELL the matrix rows might be sorted by a sorting scope called SIGMA.
The vectorization itself doesn't take of this. But, the output has to be
sorted to get the results in the right order. This patch introduces a proxy
writer, which just takes the user requested writer and sorts the output, before
calling stepFinished() on the real writer. Also the grid is sorted back for
further computations.
To do this as writer has the advantage, that the output is only sorted, when
output is done and not necessary every timestep (depends on outputFrequency).

--HG--
extra : histedit_source : 2fc2f494844c7839c84414ce3aefdf003e9664a3