Filter/Selector must specify memory location (CUDA/host)
This simple, yet big change to the interface of Filter and Selector is
a preparation to optimize IO for our CUDA backends.

Currently IO involves copying the whole grid from the device to the
host. This is grossly inefficient for multiple reasons: in most cases
a user won't need all simulation data, but just a selection of
variables. Sometimes a user might be interested in just a subset of
the grid. Finally, especially with CUDA devices, the output target
might be a GPU, so copying the grid data from the device to host
memory and back to the device is unnecessary.

A Filter is basically a specialized memcpy which will extract (or
insert) fractions (certain variables at certain coordinates) of a
grid. Every Filter now needs to be aware of the locations of source
and target memory. Grids have to specify one of these locations
(depending on the direction of data transfer), but users of a Selector
(e.g. Writers, Steerers) need to specify the other location.

The downside of this change is that ripples through to user-visible IO
components, so users who roll their own Writer/Steerer together with a
Selector will need to adapt their code.