random: removed register keyword
In newer C++ standards this leads to warnings: warning: 'register'
storage class specifier is deprecated and incompatible with C++1z
[-Wdeprecated-register]

Furthermore, this is not needed in older standards either, as the
compiler decides which variables go into registers. The register keyword
might even slow performance down, as the compiler can get confused.

This patch removes the register keyword and resolves the warnings.